requirements.txt
# Data Collection and API
spotipy==2.20.0                # Spotify API wrapper
python-dotenv==1.0.0           # Load environment variables from .env file

# Data Handling and Processing
pandas==1.5.3                  # Data manipulation and analysis
numpy==1.23.5                  # Numerical operations
scikit-learn==1.1.3            # Machine learning tools, normalization, encoding

# Deep Learning and Modeling
torch==1.13.1                  # PyTorch for building models
torchvision==0.14.1            # Optional: For any pre-trained models or utilities

# Deployment and Serving
fastapi==0.85.1                # For creating the recommendation API
uvicorn==0.18.3                # ASGI server for FastAPI

# AWS and S3 Management
boto3==1.26.3                  # AWS SDK for Python to interact with S3

# MLOps and Experiment Tracking
mlflow==2.0.1                  # Experiment tracking, model versioning
sqlalchemy==1.4.42             # Database management for MLflow

# Logging and Monitoring
loguru==0.6.0                  # Optional: For structured logging

# Visualization
matplotlib==3.6.2              # Basic plotting library for data visualization
seaborn==0.12.1                # Data visualization, works well with pandas

# Additional Utilities
joblib==1.2.0                  # Saving and loading model artifacts
